Update main.js
Modularized and organized the functions so they rely less heavily on the global variables. Was not able to eliminate global variables, but now only one function writes to them. The UI was changed significantly. Each hex has a text label on load to say what it is. The output div has been eliminated, and the canvas has been centered on the screen.